restore - simplify project structure, begin to impl restore #2
