+++ internal/common/ts.go
@@ -1,0 +1,9 @@
+package common
+
+import "time"
+
+const (
+	TimestampLayout = "20060102150405"
+)
+
+var WorkingTimestamp = time.Now().Truncate(time.Second).Format(TimestampLayout)
